# Conflicts:
#	core/src/ua/carcassone/game/screens/GameScreen.java

+++ core/src/ua/carcassone/game/game/TileType.java
@@ -1,12 +1,12 @@
 package ua.carcassone.game.game;
 
 public class TileType {
-    byte[] sides = new byte[4];
-    byte[] halves = new byte[8];
+    int[] sides = new int[4];
+    int[] halves = new int[8];
     boolean monastery;
     boolean shield;
 
-    TileType(byte[] sides, byte[] halves, boolean monastery, boolean shield) {
+    TileType(int[] sides, int[] halves, boolean monastery, boolean shield) {
         this.sides = sides;
         this.halves = halves;
         this.monastery = monastery;
